pocs/pytorch-stylegan2/requirements.txt
numpy
transformers
sentence-transformers
seaborn
torch
torchvision
matplotlib
pandas
scikit-learn
nltk
gensim
tensorflow
keras
opencv-python
fastapi
uvicorn
stylegan2_pytorch
dnnlib-util
legacy